move qbittorrent classes package
